feat: Implement password encryption

+++ src/main/kotlin/com/lapis0875/backendtask/Controller.kt
@@ -2,6 +2,7 @@ package com.lapis0875.backendtask
 
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.http.HttpStatus
+import org.springframework.security.crypto.password.PasswordEncoder
 import org.springframework.transaction.TransactionSystemException
 import org.springframework.web.bind.annotation.DeleteMapping
 import org.springframework.web.bind.annotation.PathVariable
@@ -25,6 +26,9 @@ class APIController {
     private lateinit var restTemplate: RestTemplate
 
     @Autowired
+    lateinit var passwordEncoder: PasswordEncoder
+
+    @Autowired
     private lateinit var userService: UserService
 
     @Autowired
@@ -38,7 +42,7 @@ class APIController {
     fun signup(@RequestBody req: UserRequest): UserResponse {
         // 기존에 가입된 유저가 있는지 확인
         try {
-            userService.getByPayload(req.email, req.password)
+            userService.getByPayload(req.email)
             throw ResponseStatusException(HttpStatus.BAD_REQUEST, "Cannot sign in with already existing email")       // 기존 유저 존재시, 같은 정보로의 가입을 막아야 함.
         } catch (e: EntityNotFoundException) {
             val user: UserDTO = userService.create(User(req.email, req.password, req.username!!))       // 신규 유저 생성.
@@ -50,15 +54,20 @@ class APIController {
     @ResponseStatus(HttpStatus.OK)
     fun signout(@RequestBody req: UserRequest) {
         val user: User = try {
-            userService.getByPayload(req.email, req.password)
-        } catch (e: EntityNotFoundException) { throw ResponseStatusException(HttpStatus.BAD_REQUEST) }  // 유저가 존재하지 않을 때 오류 발생.
-        userService.remove(user.userId)
+            userService.getByPayload(req.email)
+        } catch (e: EntityNotFoundException) { throw ResponseStatusException(HttpStatus.BAD_REQUEST, "Cannot signout non-existing account") }  // 유저가 존재하지 않을 때 오류 발생.
+        if (userService.compare(user, req.email, req.password))
+            userService.remove(user.userId)
+        else
+            throw ResponseStatusException(HttpStatus.BAD_REQUEST, "Invalid password")
     }
 
     @PostMapping("/article")
     fun createArticle(@RequestBody req: ArticleRequest): ArticleResponse {
         try {
-            val user: User = userService.getByPayload(req.email, req.password)
+            val user: User = userService.getByPayload(req.email)
+            if (!passwordEncoder.matches(req.password, user.password))
+                throw ResponseStatusException(HttpStatus.BAD_REQUEST, "Invalid password")
             return articleService.create(Article(content = req.content, title = req.title, user = user)).toResponse()
         } catch (e: EntityNotFoundException) {
             throw ResponseStatusException(
@@ -80,12 +89,20 @@ class APIController {
     @PutMapping("/article/{id}")
     fun editArticle(@PathVariable id: Long, @RequestBody req: ArticleRequest): ArticleResponse {
         try {
-            val user: User = userService.getByPayload(req.email, req.password)
-            return articleService.update(id, Article(content = req.content, title = req.title, articleId = id, user = user)).toResponse()
+            val article = articleService.getById(id)
+            if (userService.compare(article.user, req.email, req.password)) {
+                article.content = req.content
+                article.title = req.title
+                return articleService.update(id, article).toResponse()
+            }
+            else throw ResponseStatusException(
+                HttpStatus.BAD_REQUEST,
+                "Cannot edit article with invalid account"
+            )
         } catch (e: EntityNotFoundException) {
             throw ResponseStatusException(
                 HttpStatus.BAD_REQUEST,
-                "Cannot create article with invalid account",
+                "Cannot edit non-existing article",
                 e
             )
         }  catch (e: TransactionSystemException) {
@@ -101,21 +118,32 @@ class APIController {
 
     @DeleteMapping("/article/{id}")
     fun removeArticle(@PathVariable id: Long, @RequestBody req: UserRequest) {
-        val article: ArticleDTO = articleService.getByIdWrapped(id)
-        if (article.run {email == req.email && password == req.password})
-            articleService.remove(id)
-        else
+        try {
+            val article: Article = articleService.getById(id)
+            if (userService.compare(article.user, req.email, req.password))
+                articleService.remove(id)
+            else
+                throw ResponseStatusException(
+                    HttpStatus.BAD_REQUEST,
+                    "Only author of article can remove it",
+                )
+        }
+        catch (e: EntityNotFoundException) {
             throw ResponseStatusException(
                 HttpStatus.BAD_REQUEST,
-                "Only author of article can remove it",
+                "Cannot remove non-existing article",
+                e
             )
+        }
     }
 
     @PostMapping("/comment/{id}")
     @Throws(ResponseStatusException::class)
     fun writeComment(@PathVariable id: Long, @RequestBody req: CommentRequest): CommentResponse {
         try {
-            val author: User = userService.getByPayload(req.email, req.password)
+            val author: User = userService.getByPayload(req.email)
+            if (!passwordEncoder.matches(req.password, author.password))
+                throw ResponseStatusException(HttpStatus.BAD_REQUEST, "Invalid password")
             val article: Article = articleService.getById(id)
             return commentService.create(Comment(req.content, article = article, user = author)).toResponse()
         } catch (e: EntityNotFoundException) {
@@ -141,24 +169,28 @@ class APIController {
         @PathVariable commentId: Long,
         @RequestBody req: CommentRequest
     ): CommentResponse {
-        val article: Article = articleService.getById(articleId)
-        if (article.run { user.email != req.email || user.password != req.password }) {
-            throw ResponseStatusException(
-                HttpStatus.BAD_REQUEST,
-                "Only author of the comment can edit it"
-            )
-        }
         try {
-            return commentService.update(commentId, Comment(req.content, commentId = commentId, article = article, user = article.user)).toResponse()
-        } catch (e: TransactionSystemException) {
-            if (e.rootCause is ConstraintViolationException) {
+            val article: Article = articleService.getById(articleId)
+            val comment: Comment = commentService.getById(commentId)
+
+            if (userService.compare(comment.user, req.email, req.password)) {
                 throw ResponseStatusException(
                     HttpStatus.BAD_REQUEST,
-                    "Cannot have comment with content",
-                    e
+                    "Only author of the comment can edit it"
                 )
-            } else { throw e }
-        }
+            }
+            try {
+                return commentService.update(commentId, Comment(req.content, commentId = commentId, article = article, user = article.user)).toResponse()
+            } catch (e: TransactionSystemException) {
+                if (e.rootCause is ConstraintViolationException) {
+                    throw ResponseStatusException(
+                        HttpStatus.BAD_REQUEST,
+                        "Cannot have comment with content",
+                        e
+                    )
+                } else { throw e }
+            }
+        } catch (e: EntityNotFoundException) {throw ResponseStatusException(HttpStatus.BAD_REQUEST, "Article or Comment does not exist")}
     }
 
     @DeleteMapping("/comment/{articleId}/{commentId}")
@@ -168,13 +200,14 @@ class APIController {
         @PathVariable commentId: Long,
         @RequestBody req: UserRequest
     ) {
-        val article: ArticleDTO = articleService.getByIdWrapped(articleId)
-        val comment: CommentDTO = commentService.getByIdWrapped(commentId)
-        if (comment.commentId == commentId && article.run {
-            this.articleId == articleId
-            && this.email == req.email
-            && this.password == req.password
-        }) return commentService.remove(commentId)
+        val comment: Comment = commentService.getById(commentId)
+        if (comment.article.articleId != articleId)
+            throw ResponseStatusException(
+                HttpStatus.BAD_REQUEST,
+                "Invalid article id for this comment"
+            )
+        if (userService.compare(comment.user, req.email, req.password) && comment.article.articleId == articleId)
+            return commentService.remove(commentId)
         else
             throw ResponseStatusException(
                 HttpStatus.BAD_REQUEST,
